migrate useSelfHostedUserList to SelfHostedUserListQueryOpts
