+++ src/services/selfHosted/SelfHostedUserListQueryOpts.ts
@@ -1,9 +1,8 @@
-import { useInfiniteQuery } from '@tanstack/react-query'
-import { useMemo } from 'react'
+import { infiniteQueryOptions as infiniteQueryOptionsV5 } from '@tanstack/react-queryV5'
 import { z } from 'zod'
 
 import Api from 'shared/api'
-import { NetworkErrorObject } from 'shared/api/helpers'
+import { rejectNetworkError } from 'shared/api/helpers'
 
 const OwnerSchema = z.object({
   ownerid: z.number(),
@@ -24,20 +23,20 @@ const RequestSchema = z.object({
   totalPages: z.number(),
 })
 
-interface UseSelfHostedUserListArgs {
+interface SelfHostedUserListQueryArgs {
   activated?: boolean
   search?: string
   isAdmin?: boolean
 }
 
-export const useSelfHostedUserList = ({
+export const SelfHostedUserListQueryOpts = ({
   activated,
   search,
   isAdmin,
-}: UseSelfHostedUserListArgs) => {
-  const { data, ...rest } = useInfiniteQuery({
+}: SelfHostedUserListQueryArgs) => {
+  return infiniteQueryOptionsV5({
     queryKey: ['SelfHostedUserList', activated, search, isAdmin],
-    queryFn: ({ pageParam = 1, signal }) => {
+    queryFn: ({ pageParam, signal }) => {
       return Api.get({
         path: '/users',
         signal,
@@ -51,32 +50,24 @@ export const useSelfHostedUserList = ({
         const parsedData = RequestSchema.safeParse(res)
 
         if (!parsedData.success) {
-          return Promise.reject({
+          return rejectNetworkError({
             status: 404,
             data: {},
-            dev: 'useSelfHostedUserList - 404 schema parsing failed',
-          } satisfies NetworkErrorObject)
+            dev: 'SelfHostedUserListQueryOpts - 404 schema parsing failed',
+            error: parsedData.error,
+          })
         }
+
         return parsedData.data
       })
     },
-    suspense: false,
+    initialPageParam: '1',
     getNextPageParam: (data) => {
       if (data?.next) {
         const { searchParams } = new URL(data.next)
         return searchParams.get('page')
       }
-      return undefined
+      return null
     },
   })
-
-  const memoedData = useMemo(
-    () => data?.pages?.flatMap((page) => page.results) ?? [],
-    [data]
-  )
-
-  return {
-    ...rest,
-    data: memoedData,
-  }
 }
