+++ src/services/selfHosted/index.ts
@@ -1,5 +1,4 @@
 export * from './useSelfHostedCurrentUser'
 export * from './useSelfHostedSeatsConfig'
-export * from './useSelfHostedUserList'
 export * from './useSelfHostedHasAdmins'
 export * from './useSelfHostedSeatsAndLicense'
